requirements.txt
Django>=5.2
djangorestframework
psycopg2-binary
pandas
openpyxl

